fixing some formatting and organization
